+++ OCaml/Basics.v
@@ -34,6 +34,8 @@ Parameter exn : Set.
 
 Parameter extensible_type : Set.
 
+Parameter extensible_type_value : extensible_type.
+
 Definition int32 := Z.
 
 Definition int64 := Z.
@@ -437,6 +439,9 @@ Module CamlinternalFormatBasics.
 
   Parameter Format : forall {A : Set}, Any -> string -> A.
   Parameter Int32 : Any -> Any -> Any -> Any -> Any.
+  Parameter Int64 : Any -> Any -> Any -> Any -> Any.
+  Parameter String_literal : string -> Any -> Any.
+  Parameter String : Any -> Any -> Any.
 
   Parameter End_of_format : Any.
   Parameter Int_d : Any.